requirements.txt
schedule==1.2.2
clickhouse-driver==0.2.9
geoip2==5.1.0
pandas==2.2.3
clickhouse-connect==0.8.18
python-dotenv==1.1.1
netmiko==4.6.0
jinja2==3.1.6
napalm==5.0.0